Finishing adding in the pagination
