+++ gulp/src/analytics/components/Table/TableContainer.js
@@ -15,57 +15,59 @@ class TableContainer extends Component {
     };
 
     this.handlePageSizeChange = this.handlePageSizeChange.bind(this);
-    this.pager = this.pager.bind(this);
+    this.handlePageChange = this.handlePageChange.bind(this);
+    // this.pager = this.pager.bind(this);
   }
   componentWillReceiveProps() {
     this.setState({
       currentPage: 1,
     });
   }
-  getNumPages() {
-    const {tableData} = this.props;
-    const pageSize = this.state.pageSize;
-    const paginationData = tableData.PageLoadData.rows;
-    let numPages = Math.floor(paginationData.length / pageSize);
-    if (paginationData.length % pageSize > 0) {
-      numPages++
-    }
-    return numPages;
-  }
+  // getNumPages() {
+  //   const {tableData} = this.props;
+  //   const pageSize = this.state.pageSize;
+  //   const paginationData = tableData.PageLoadData.rows;
+  //   let numPages = Math.floor(paginationData.length / pageSize);
+  //   if (paginationData.length % pageSize > 0) {
+  //     numPages++
+  //   }
+  //   return numPages;
+  // }
   handlePageChange(pageNum) {
     this.setState({
       currentPage: pageNum
     });
+    console.log(pageNum);
   }
   handlePageSizeChange(e) {
     this.setState({
       pageSize: Number(e.target.value),
     });
   }
-  pager() {
-    const currentPage = this.state.currentPage;
-    const numPages = this.getNumPages();
-    const handleClick = (pageNum) => {return this.handlePageChange(pageNum)};
-    const pageLinks = [];
-    if (currentPage > 1) {
-       if (currentPage > 2) {
-         pageLinks.push(<span className="pageLink" onClick={handleClick(1)}>«</span>);
-         pageLinks.push(' ');
-       }
-       pageLinks.push(<span className="pageLink" onClick={handleClick(currentPage - 1)}>‹</span>);
-       pageLinks.push(' ');
-     }
-     pageLinks.push(<span className="currentPage">Page {currentPage} of {numPages}</span>);
-     if (currentPage < numPages) {
-       pageLinks.push(' ');
-       pageLinks.push(<span className="pageLink" onClick={handleClick(currentPage + 1)}>›</span>);
-       if (currentPage < numPages - 1) {
-         pageLinks.push(' ');
-         pageLinks.push(<span className="pageLink" onClick={handleClick(numPages)}>»</span>);
-       }
-     }
-     return <div className="pager">{pageLinks}</div>;
-   }
+  // pager() {
+  //   const currentPage = this.state.currentPage;
+  //   const numPages = this.getNumPages();
+  //   const handleClick = (pageNum) => {return this.handlePageChange(pageNum)};
+  //   const pageLinks = [];
+  //   if (currentPage > 1) {
+  //      if (currentPage > 2) {
+  //        pageLinks.push(<span className="pageLink" onClick={handleClick(1)}>«</span>);
+  //        pageLinks.push(' ');
+  //      }
+  //      pageLinks.push(<span className="pageLink" onClick={handleClick(currentPage - 1)}>‹</span>);
+  //      pageLinks.push(' ');
+  //    }
+  //    pageLinks.push(<span className="currentPage">Page {currentPage} of {numPages}</span>);
+  //    if (currentPage < numPages) {
+  //      pageLinks.push(' ');
+  //      pageLinks.push(<span className="pageLink" onClick={handleClick(currentPage + 1)}>›</span>);
+  //      if (currentPage < numPages - 1) {
+  //        pageLinks.push(' ');
+  //        pageLinks.push(<span className="pageLink" onClick={handleClick(numPages)}>»</span>);
+  //      }
+  //    }
+  //    return <div className="pager">{pageLinks}</div>;
+  //  }
   render() {
     const {tableData} = this.props;
     const pageSize = this.state.pageSize;
@@ -76,7 +78,7 @@ class TableContainer extends Component {
     const paginationData = rowData.slice(start, end);
     return (
       <div>
-        <Table enablePaging={this.pager} handlePageSizeChange={this.handlePageSizeChange} pageSize={this.state.pageSize} tableData={tableData}>
+        <Table pageChange={this.handlePageChange} currentPage={this.state.currentPage} handlePageSizeChange={this.handlePageSizeChange} pageSize={this.state.pageSize} tableData={tableData}>
           <TableColumns columnData={tableData}/>
           <TableRows pageSize={this.state.pageSize} rowData={paginationData} tableData={tableData} />
         </Table>
